Start basic work on tunneling protocol and move some code into a common crate
