+++ apps/labrinth/Cargo.toml
@@ -37,6 +37,7 @@ hyper = { version = "0.14", features = ["full"] }
 hyper-tls = "0.5.0"
 
 serde_json = "1.0"
+serde_cbor = "0.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_with = "3.0.0"
 chrono = { version = "0.4.26", features = ["serde"] }
@@ -74,6 +75,7 @@ dotenvy = "0.15.7"
 log = "0.4.20"
 env_logger = "0.10.1"
 thiserror = "1.0.56"
+either = "1.13"
 
 sqlx = { version = "0.8.2", features = [
     "runtime-tokio-rustls",
@@ -123,6 +125,8 @@ jemallocator = { version = "0.5.4", optional = true }
 async-stripe = { version = "0.39.1", features = ["runtime-tokio-hyper-rustls"] }
 rusty-money = "0.4.1"
 json-patch = "*"
+
+rust-common = { path = "../../packages/rust-common" }
 
 [dev-dependencies]
 actix-http = "3.4.0"
