+++ apps/labrinth/src/auth/oauth/errors.rs
@@ -1,9 +1,9 @@
 use super::ValidatedRedirectUri;
 use crate::auth::AuthenticationError;
 use crate::models::error::ApiError;
-use crate::models::ids::DecodingError;
 use actix_web::http::{header::LOCATION, StatusCode};
 use actix_web::HttpResponse;
+use rust_common::ids::DecodingError;
 
 #[derive(thiserror::Error, Debug)]
 #[error("{}", .error_type)]
