+++ apps/labrinth/src/routes/internal/moderation.rs
@@ -1,12 +1,12 @@
 use super::ApiError;
 use crate::database;
 use crate::database::redis::RedisPool;
-use crate::models::ids::random_base62;
 use crate::models::projects::ProjectStatus;
 use crate::queue::moderation::{ApprovalType, IdentifiedFile, MissingMetadata};
 use crate::queue::session::AuthQueue;
 use crate::{auth::check_is_moderator_from_headers, models::pats::Scopes};
 use actix_web::{web, HttpRequest, HttpResponse};
+use rust_common::ids::random_base62;
 use serde::Deserialize;
 use sqlx::PgPool;
 use std::collections::HashMap;
