+++ apps/labrinth/src/routes/v3/project_creation.rs
@@ -8,7 +8,6 @@ use crate::database::models::{self, image_item, User};
 use crate::database::redis::RedisPool;
 use crate::file_hosting::{FileHost, FileHostingError};
 use crate::models::error::ApiError;
-use crate::models::ids::base62_impl::to_base62;
 use crate::models::ids::{ImageId, OrganizationId};
 use crate::models::images::{Image, ImageContext};
 use crate::models::pats::Scopes;
@@ -32,6 +31,7 @@ use chrono::Utc;
 use futures::stream::StreamExt;
 use image::ImageError;
 use itertools::Itertools;
+use rust_common::ids::base62_impl::to_base62;
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::postgres::PgPool;
