+++ apps/labrinth/src/routes/v3/projects.rs
@@ -11,7 +11,6 @@ use crate::database::redis::RedisPool;
 use crate::database::{self, models as db_models};
 use crate::file_hosting::FileHost;
 use crate::models;
-use crate::models::ids::base62_impl::parse_base62;
 use crate::models::images::ImageContext;
 use crate::models::notifications::NotificationBody;
 use crate::models::pats::Scopes;
@@ -33,6 +32,7 @@ use actix_web::{web, HttpRequest, HttpResponse};
 use chrono::Utc;
 use futures::TryStreamExt;
 use itertools::Itertools;
+use rust_common::ids::base62_impl::parse_base62;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use sqlx::PgPool;
