+++ apps/labrinth/src/routes/v3/reports.rs
@@ -6,9 +6,7 @@ use crate::database::models::thread_item::{
 };
 use crate::database::redis::RedisPool;
 use crate::models::ids::ImageId;
-use crate::models::ids::{
-    base62_impl::parse_base62, ProjectId, UserId, VersionId,
-};
+use crate::models::ids::{ProjectId, UserId, VersionId};
 use crate::models::images::{Image, ImageContext};
 use crate::models::pats::Scopes;
 use crate::models::reports::{ItemType, Report};
@@ -19,6 +17,7 @@ use crate::util::img;
 use actix_web::{web, HttpRequest, HttpResponse};
 use chrono::Utc;
 use futures::StreamExt;
+use rust_common::ids::base62_impl::parse_base62;
 use serde::Deserialize;
 use sqlx::PgPool;
 use validator::Validate;
