+++ apps/labrinth/tests/project.rs
@@ -12,10 +12,10 @@ use futures::StreamExt;
 use labrinth::database::models::project_item::{
     PROJECTS_NAMESPACE, PROJECTS_SLUGS_NAMESPACE,
 };
-use labrinth::models::ids::base62_impl::parse_base62;
 use labrinth::models::projects::ProjectId;
 use labrinth::models::teams::ProjectPermissions;
 use labrinth::util::actix::{MultipartSegment, MultipartSegmentData};
+use rust_common::ids::base62_impl::parse_base62;
 use serde_json::json;
 
 use crate::common::api_common::models::CommonProject;
