+++ apps/labrinth/tests/scopes.rs
@@ -18,10 +18,10 @@ use common::environment::{
     with_test_environment, with_test_environment_all, TestEnvironment,
 };
 use common::{database::*, scopes::ScopeTest};
-use labrinth::models::ids::base62_impl::parse_base62;
 use labrinth::models::pats::Scopes;
 use labrinth::models::projects::ProjectId;
 use labrinth::models::users::UserId;
+use rust_common::ids::base62_impl::parse_base62;
 use serde_json::json;
 
 // For each scope, we (using test_scope):
