+++ apps/labrinth/tests/v2/project.rs
@@ -20,12 +20,10 @@ use futures::StreamExt;
 use itertools::Itertools;
 use labrinth::{
     database::models::project_item::PROJECTS_SLUGS_NAMESPACE,
-    models::{
-        ids::base62_impl::parse_base62, projects::ProjectId,
-        teams::ProjectPermissions,
-    },
+    models::{projects::ProjectId, teams::ProjectPermissions},
     util::actix::{AppendsMultipart, MultipartSegment, MultipartSegmentData},
 };
+use rust_common::ids::base62_impl::parse_base62;
 use serde_json::json;
 
 #[actix_rt::test]
