+++ apps/labrinth/tests/v2/scopes.rs
@@ -6,9 +6,9 @@ use crate::common::dummy_data::TestFile;
 use crate::common::environment::with_test_environment;
 use crate::common::environment::TestEnvironment;
 use crate::common::scopes::ScopeTest;
-use labrinth::models::ids::base62_impl::parse_base62;
 use labrinth::models::pats::Scopes;
 use labrinth::models::projects::ProjectId;
+use rust_common::ids::base62_impl::parse_base62;
 
 // Project version creation scopes
 #[actix_rt::test]
