+++ apps/labrinth/tests/v2/search.rs
@@ -11,7 +11,7 @@ use crate::common::environment::with_test_environment;
 use crate::common::environment::TestEnvironment;
 use actix_http::StatusCode;
 use futures::stream::StreamExt;
-use labrinth::models::ids::base62_impl::parse_base62;
+use rust_common::ids::base62_impl::parse_base62;
 use serde_json::json;
 use std::collections::HashMap;
 use std::sync::Arc;
