+++ packages/app-lib/Cargo.toml
@@ -29,6 +29,7 @@ regex = "1.5"
 sys-info = "0.9.0"
 sysinfo = "0.30.8"
 thiserror = "1.0"
+either = "1.13"
 
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.18", features = ["chrono", "env-filter"] }
@@ -61,6 +62,8 @@ byteorder = "1.5.0"
 base64 = "0.22.0"
 
 sqlx = { version = "0.8.2", features = [ "runtime-tokio", "sqlite", "macros" ] }
+
+rust-common = { path = "../rust-common" }
 
 [target.'cfg(windows)'.dependencies]
 winreg = "0.52.0"
