+++ packages/app-lib/src/api/mod.rs
@@ -19,8 +19,9 @@ pub mod data {
         Hooks, JavaVersion, LinkedData, MemorySettings, ModLoader,
         ModrinthCredentials, Organization, ProcessMetadata, ProfileFile,
         Project, ProjectType, SearchResult, SearchResults, Settings,
-        TeamMember, Theme, User, UserFriend, UserStatus, Version, WindowSize,
+        TeamMember, Theme, User, UserFriend, Version, WindowSize,
     };
+    pub use rust_common::users::UserStatus;
 }
 
 pub mod prelude {
