+++ packages/rust-common/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "rust-common"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_bytes = "0.11"
+serde_json = "1.0"
+serde_cbor = "0.11"
+
+uuid = { version = "1.12", features = ["serde"] }
+chrono = { version = "0.4", features = ["serde"] }
+
+actix-ws = "0.3"
+
+thiserror = "2.0"
+rand = "0.8"
+either = "1.13"
