+++ packages/rust-common/src/networking/mod.rs
@@ -1,0 +1,2 @@
+pub mod message;
+pub mod wire;
